apps/bot/requirements.txt
requests
python-dotenv
python-telegram-bot[job-queue]
SQLAlchemy==2.0.1
psycopg2-binary